refactor: [FE] 반복 사용되는 소켓 이벤트 리팩터링 useUserSocket에 emitUpdateUserStream를 만들어서 여러곳에서 재사용될 수 있도록 변경했어요.
